added fixture for checking operator
