+++ tests/functional-tests/ibm_spectrum_scale_csi/scale_operator.py
@@ -862,6 +862,37 @@ def read_operator_data(clusterconfig, namespace, kubeconfig=None):
 
     return data
 
+def get_remote_data(data_passed):
+    remote_data = copy.deepcopy(data_passed)
+    remote_data["remoteFs_remote_name"] = ff.get_remoteFs_remotename(copy.deepcopy(remote_data))
+    if remote_data["remoteFs_remote_name"] is None:
+        LOGGER.error("Unable to get remoteFs , name on remote cluster")
+        assert False
+
+    remote_data["primaryFs"] = remote_data["remoteFs_remote_name"]
+    remote_data["id"] = remote_data["remoteid"]
+    remote_data["port"] = remote_data["remote_port"]
+    for cluster in remote_data["clusters"]:
+        if cluster["id"] == remote_data["remoteid"]:
+            remote_data["guiHost"] = cluster["restApi"][0]["guiHost"]
+            remote_sec_name = cluster["secrets"]
+            remote_data["username"] = remote_data["remote_username"][remote_sec_name]
+            remote_data["password"] = remote_data["remote_password"][remote_sec_name]
+
+    remote_data["volDirBasePath"] = remote_data["r_volDirBasePath"]
+    remote_data["parentFileset"] = remote_data["r_parentFileset"]
+    remote_data["gid_name"] = remote_data["r_gid_name"]
+    remote_data["uid_name"] = remote_data["r_uid_name"]
+    remote_data["gid_number"] = remote_data["r_gid_number"]
+    remote_data["uid_number"] = remote_data["r_uid_number"]
+    remote_data["inodeLimit"] = remote_data["r_inodeLimit"]
+    # for get_mount_point function
+    remote_data["type_remote"] = {"username": data_passed["username"],
+                                  "password": data_passed["password"],
+                                  "port": data_passed["port"],
+                                  "guiHost": data_passed["guiHost"]}
+
+    return remote_data
 
 def get_cmd_values(request):
     kubeconfig_value = request.config.option.kubeconfig
