+++ tests/functional-tests/tests/volume_snapshot.py
@@ -6,38 +6,15 @@ LOGGER = logging.getLogger()
 pytestmark = [pytest.mark.volumesnapshot, pytest.mark.localcluster]
 
 @pytest.fixture(scope='session', autouse=True)
-def values(request):
+def values(request, check_csi_operator):
     global data, snapshot_object, kubeconfig_value  # are required in every testcase
     kubeconfig_value, clusterconfig_value, operator_namespace, test_namespace, runslow_val, operator_yaml = scaleop.get_cmd_values(request)
 
     data = scaleop.read_driver_data(clusterconfig_value, test_namespace, operator_namespace, kubeconfig_value)
-    operator_data = scaleop.read_operator_data(clusterconfig_value, operator_namespace, kubeconfig_value)
     keep_objects = data["keepobjects"]
-    ff.cred_check(data)
-    ff.set_data(data)
+    if not(data["volBackendFs"] == ""):
+        data["primaryFs"] = data["volBackendFs"]
 
-    operator = scaleop.Scaleoperator(kubeconfig_value, operator_namespace, operator_yaml)
-    operator_object = scaleop.Scaleoperatorobject(operator_data, kubeconfig_value)
-    condition = scaleop.check_ns_exists(kubeconfig_value, operator_namespace)
-    if condition is True:
-        if not(operator_object.check(data["csiscaleoperator_name"])):
-            LOGGER.error("Operator custom object is not deployed succesfully")
-            assert False
-    else:
-        operator.create()
-        operator.check()
-        scaleop.check_nodes_available(operator_data["pluginNodeSelector"], "pluginNodeSelector")
-        scaleop.check_nodes_available(
-            operator_data["provisionerNodeSelector"], "provisionerNodeSelector")
-        scaleop.check_nodes_available(
-            operator_data["attacherNodeSelector"], "attacherNodeSelector")
-        operator_object.create()
-        val = operator_object.check()
-        if val is True:
-            LOGGER.info("Operator custom object is deployed succesfully")
-        else:
-            LOGGER.error("Operator custom object is not deployed succesfully")
-            assert False
     if runslow_val:
         value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"},
                      {"access_modes": "ReadWriteOnce", "storage": "1Gi"}]
@@ -48,16 +25,6 @@ def values(request):
     number_of_snapshots = 1
     snapshot_object = scaleop.Snapshot(kubeconfig_value, test_namespace, keep_objects, value_pvc, value_vs_class,
                                        number_of_snapshots, data["image_name"], data["id"], data["pluginNodeSelector"])
-    if not(data["volBackendFs"] == ""):
-        data["primaryFs"] = data["volBackendFs"]
-    ff.create_dir(data["volDirBasePath"])
-
-    yield
-    if condition is False and not(keep_objects):
-        operator_object.delete()
-        operator.delete()
-        if(ff.fileset_exists(data)):
-            ff.delete_fileset(data)
 
 
 @pytest.mark.regression
